requirements.txt
Django==4.2.5
psycopg2-binary==2.9.9
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
django-filter==23.3
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0









